more rearranging to allow building platforms independently
